Fix clang-tidy errors in the linux-emulator code
Apply tidy-driven fixes across the new linux-emulator subsystem:
- Use designated initializers for aggregate construction (mapped_region,
  auxv_entry, cached_dir_entry, linux_syscall_handler_entry, library_info,
  vdso_func, fixture_expectation, ...).
- Make ELF / kernel mirror structs and fixed-size scratch buffers used
  for snprintf opt out of the C-array checks via NOLINTBEGIN/END since
  they intentionally mirror the kernel ABI layout.
- Switch logger calls and printf-style logging to PRIx64/PRIu64 from
  <cinttypes> so the format matches uint64_t portably; drop the
  unsigned-long-long casts that triggered google-runtime-int.
- Mark stateless procfs / signal_dispatcher / linux_fd_table / file_system
  helpers static and update call sites accordingly.
- Use std::max / std::ranges algorithms / contains() / starts_with() /
  proper pass-by-value-and-move where the linter pointed it out.
- Reserve vectors before push_back loops in process context setup.
- Brace single-statement bodies.
- Add cstring include + drop the gcc 13 false-positive
  lexicographical_compare path in dns_resolver (kept from prior commit).

The release preset still passes; tidy now reports zero errors.

+++ src/linux-analyzer/main.cpp
@@ -1,5 +1,6 @@
 #include <cstdio>
 #include <cstdlib>
+#include <cinttypes>
 #include <string>
 #include <vector>
 #include <filesystem>
@@ -37,6 +38,7 @@ namespace
         fprintf(stderr, "  --help           Show this help message\n");
     }
 
+    // NOLINTNEXTLINE(cppcoreguidelines-avoid-c-arrays,hicpp-avoid-c-arrays,modernize-avoid-c-arrays)
     cli_options parse_args(const int argc, char* argv[])
     {
         cli_options opts{};
@@ -154,7 +156,7 @@ int main(const int argc, char* argv[])
 
         printf("\n--- Emulation finished ---\n");
         printf("Exit status: %d\n", final_exit_status);
-        printf("Instructions executed: %llu\n", static_cast<unsigned long long>(instructions));
+        printf("Instructions executed: %" PRIu64 "\n", instructions);
 
         return final_exit_status;
     }
